refactor: extract pipeline-safe deploy steps into src/pipeline/ (GitOps PR1 step 1)
Behavior-neutral extraction per plans/gitops-deployment-strategies.md section 6
PR1: the deploy steps that emit will run in CI move out of the 4300-line
src/cli/deploy.ts into src/pipeline/, so emit and deploy call the SAME
functions and can never disagree on validation.

Module map (incident comments moved with their code):
- src/pipeline/images.ts (A6/A7 + L15): refreshFetchCacheStaging,
  buildDockerCommands, assertSafePoolName.
- src/pipeline/digests.ts (A8): resolveDeployImageDigests,
  resolveRegistryDigestAny, resolveRegistryDigest, resolveImageDigest,
  DIGEST_RE — the S7/S23/S25/S27/S28 fail-closed registry-first rules and
  --allow-mutable-tags opt-out, verbatim.
- src/pipeline/fingerprints.ts (A2 + B2): the inline runDeploy blocks became
  resolveBuiltTargetPlatform, assertTargetFingerprint,
  assertDeployablePoolTopology, assertBuildIdChangedSinceServing (N14),
  assertNoCrossBuildNameCollision, assertNoSelfNameCollision (N62). Error
  messages are byte-identical; guard ordering in runDeploy is unchanged
  (N14 before the composed-name guard, N62 after the cross-build guard).

deploy.ts re-exports every moved public symbol, so its import surface is
unchanged for src/cli/index.ts, tests, and any external consumer.

Regression net: the unit tests for the moved functions moved to
tests/pipeline/ (images 19, digests 36 — all 112 pre-move tests accounted
for across the split); the runDeploy orchestration tests, which exercise
the A2/B2 guards and the A6-A8 flow end to end through the CLI module,
stay in tests/cli/ untouched and pass against the re-exported surface.

Gates: npx tsc --noEmit, npx oxlint, npx oxfmt --check, npx vitest run
(129 files, 2601 passed / 19 skipped).
